+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * ruby-shadow is not installed on cygwin platform anymore. (CHEF-4946)
 * Upgrade chef-zero to 2.0, remove native-compiled puma as chef dependency. (CHEF-4901/CHEF-5005)
 * Don't honor splay when sent USR1 signal.
+* Don't set log_level in client.rb by default (CHEF-3698)
 
 ## Last Release: 11.10.0 (02/06/2014)